extract wrapper model as utility
