+++ runtimes/alibi-explain/tests/test_alibi_runtime_base.py
@@ -8,7 +8,7 @@ import pytest
 from alibi.api.interfaces import Explanation
 from numpy.testing import assert_array_equal
 
-from mlserver import ModelSettings, MLModel
+from mlserver import ModelSettings
 from mlserver.codecs import NumpyCodec
 from mlserver.types import (
     InferenceRequest,
@@ -16,7 +16,6 @@ from mlserver.types import (
     Parameters,
     RequestInput,
     ResponseOutput,
-    MetadataTensor,
 )
 from mlserver_alibi_explain.common import (
     convert_from_bytes,
@@ -121,67 +120,6 @@ async def test_remote_predict__smoke(custom_runtime_tf, rest_client):
             ssl_verify_path="",
         )
         assert isinstance(res, InferenceResponse)
-
-
-async def test_alibi_runtime_wrapper(custom_runtime_tf: MLModel):
-    """
-    Checks that the wrapper returns back the expected valued from the underlying rt
-    """
-
-    class _MockInit(AlibiExplainRuntime):
-        def __init__(self, settings: ModelSettings):
-            self._rt = custom_runtime_tf
-
-    data = np.random.randn(10, 28, 28, 1) * 255
-    inference_request = InferenceRequest(
-        parameters=Parameters(content_type=NumpyCodec.ContentType),
-        inputs=[
-            RequestInput(
-                name="predict",
-                shape=data.shape,
-                data=data.tolist(),
-                datatype="FP32",
-            )
-        ],
-    )
-
-    # settings object is dummy and discarded
-    wrapper = _MockInit(ModelSettings(name="foo", implementation=AlibiExplainRuntime))
-
-    assert wrapper.settings == custom_runtime_tf.settings
-    assert wrapper.name == custom_runtime_tf.name
-    assert wrapper.version == custom_runtime_tf.version
-    assert wrapper.inputs == custom_runtime_tf.inputs
-    assert wrapper.outputs == custom_runtime_tf.outputs
-    assert wrapper.ready == custom_runtime_tf.ready
-
-    assert await wrapper.metadata() == await custom_runtime_tf.metadata()
-    assert await wrapper.predict(inference_request) == await custom_runtime_tf.predict(
-        inference_request
-    )
-
-    # check setters
-    dummy_shape_metadata = [
-        MetadataTensor(
-            name="dummy",
-            datatype="FP32",
-            shape=[1, 2],
-        )
-    ]
-    wrapper.inputs = dummy_shape_metadata
-    custom_runtime_tf.inputs = dummy_shape_metadata
-    assert wrapper.inputs == custom_runtime_tf.inputs
-
-    wrapper.outputs = dummy_shape_metadata
-    custom_runtime_tf.outputs = dummy_shape_metadata
-    assert wrapper.outputs == custom_runtime_tf.outputs
-
-    wrapper_public_funcs = list(filter(lambda x: not x.startswith("_"), dir(wrapper)))
-    expected_public_funcs = list(
-        filter(lambda x: not x.startswith("_"), dir(custom_runtime_tf))
-    )
-
-    assert wrapper_public_funcs == expected_public_funcs
 
 
 def fake_predictor(x):
